feat(angular): integrate `MaskitoCVA` to `MaskitoDirective`
Now the `MaskitoDirective` directive can perform the functions of the
`MaskitoCVA` directive.

This change will allow us to avoid the situation where the developer forgot to
import `MaskitoCVA`, now it just works using `MaskitoDirective`.

`MaskitoCVA` was left as a placeholder to avoid a breaking change.
