+++ projects/angular/src/index.ts
@@ -1,4 +1,3 @@
 export * from './lib/deprecated';
-export * from './lib/maskito.cva';
 export * from './lib/maskito.directive';
 export * from './lib/maskito.pipe';
