+++ projects/angular/src/lib/deprecated.ts
@@ -1,11 +1,23 @@
-import {NgModule} from '@angular/core';
+import {Directive, Input, NgModule} from '@angular/core';
+import {MaskitoOptions} from '@maskito/core';
 
-import {MaskitoCVA} from './maskito.cva';
 import {MaskitoDirective} from './maskito.directive';
 import {MaskitoPipe} from './maskito.pipe';
 
 /**
- * @deprecated Use standalone `MaskitoDirective`, `MaskitoCVA` and `MaskitoPipe` instead.
+ * @deprecated Just use `MaskitoDirective`, `MaskitoCVA` no longer needed.
+ */
+@Directive({
+    standalone: true,
+    selector: 'input[maskito], textarea[maskito]',
+})
+export class MaskitoCVA {
+    @Input()
+    maskito?: MaskitoOptions | null;
+}
+
+/**
+ * @deprecated Use standalone `MaskitoDirective` and `MaskitoPipe` instead.
  * Learn more: https://maskito.dev/frameworks/angular
  * ___
  * TODO: Delete it in v3.0 (after Taiga UI 4.0 will be released and bumped in this repository).
