+++ projects/angular/src/lib/maskito.directive.ts
@@ -6,13 +6,17 @@ import {
     NgZone,
     OnChanges,
     OnDestroy,
+    Optional,
+    Self,
 } from '@angular/core';
+import {DefaultValueAccessor} from '@angular/forms';
 import {
     Maskito,
     MASKITO_DEFAULT_ELEMENT_PREDICATE,
     MASKITO_DEFAULT_OPTIONS,
     MaskitoElementPredicate,
     MaskitoOptions,
+    maskitoTransform,
 } from '@maskito/core';
 
 @Directive({standalone: true, selector: '[maskito]'})
@@ -20,7 +24,7 @@ export class MaskitoDirective implements OnDestroy, OnChanges {
     private maskedElement: Maskito | null = null;
 
     @Input()
-    maskito: MaskitoOptions | null = MASKITO_DEFAULT_OPTIONS;
+    maskito: MaskitoOptions | null = null;
 
     @Input()
     maskitoElement: MaskitoElementPredicate = MASKITO_DEFAULT_ELEMENT_PREDICATE;
@@ -28,7 +32,23 @@ export class MaskitoDirective implements OnDestroy, OnChanges {
     constructor(
         @Inject(NgZone) private readonly ngZone: NgZone,
         @Inject(ElementRef) private readonly elementRef: ElementRef<HTMLElement>,
-    ) {}
+        @Inject(DefaultValueAccessor)
+        @Self()
+        @Optional()
+        accessor: DefaultValueAccessor | null,
+    ) {
+        if (accessor !== null) {
+            const original = accessor.writeValue.bind(accessor);
+
+            accessor.writeValue = (value: unknown) => {
+                original(maskitoTransform(String(value ?? ''), this.options));
+            };
+        }
+    }
+
+    private get options(): MaskitoOptions {
+        return this.maskito ?? MASKITO_DEFAULT_OPTIONS;
+    }
 
     async ngOnChanges(): Promise<void> {
         this.maskedElement?.destroy();
@@ -43,10 +63,7 @@ export class MaskitoDirective implements OnDestroy, OnChanges {
         }
 
         this.ngZone.runOutsideAngular(() => {
-            this.maskedElement = new Maskito(
-                predicateResult,
-                this.maskito ?? MASKITO_DEFAULT_OPTIONS,
-            );
+            this.maskedElement = new Maskito(predicateResult, this.options);
         });
     }
