+++ projects/demo/src/pages/frameworks/angular/angular-doc.component.ts
@@ -7,7 +7,7 @@ import {TuiLinkModule, TuiNotificationModule} from '@taiga-ui/core';
 
 import {NestedDocExample1} from './examples/1-nested/component';
 import {NestedDocExample2} from './examples/2-nested/component';
-import {CvaDocExample3} from './examples/3-cva/component';
+import {ProgrammaticallyDocExample3} from './examples/3-programmatically/component';
 import {PipeDocExample4} from './examples/4-pipe/component';
 
 @Component({
@@ -21,7 +21,7 @@ import {PipeDocExample4} from './examples/4-pipe/component';
         TuiLinkModule,
         NestedDocExample1,
         NestedDocExample2,
-        CvaDocExample3,
+        ProgrammaticallyDocExample3,
         PipeDocExample4,
     ],
     templateUrl: './angular-doc.template.html',
@@ -45,9 +45,9 @@ export class AngularDocPageComponent {
         Custom: import('./examples/2-nested/template.html?raw'),
     };
 
-    readonly cvaExample: TuiDocExample = {
-        TypeScript: import('./examples/3-cva/component.ts?raw'),
-        HTML: import('./examples/3-cva/template.html?raw'),
+    readonly programmaticallyExample: TuiDocExample = {
+        TypeScript: import('./examples/3-programmatically/component.ts?raw'),
+        HTML: import('./examples/3-programmatically/template.html?raw'),
     };
 
     readonly pipeExample: TuiDocExample = {
