+++ projects/demo/src/pages/frameworks/angular/angular-doc.template.html
@@ -110,17 +110,17 @@
         </tui-doc-example>
 
         <tui-doc-example
-            id="cva"
-            heading="CVA"
-            [content]="cvaExample"
-            [description]="cvaDescription"
+            id="programmatically"
+            heading="Set value programmatically"
+            [content]="programmaticallyExample"
+            [description]="programmaticallyDescription"
         >
-            <ng-template #cvaDescription>
+            <ng-template #programmaticallyDescription>
                 When directly on native input/textarea tag,
-                <code>MaskitoCVA</code>
+                <code>MaskitoDirective</code>
                 formats value set programmatically with Angular forms.
             </ng-template>
-            <cva-doc-example-3></cva-doc-example-3>
+            <programmatically-doc-example-3></programmatically-doc-example-3>
         </tui-doc-example>
 
         <tui-doc-example
